Begin adding generic CI providers
