+++ SeaEye/models/ProviderManager.swift
@@ -1,0 +1,20 @@
+//
+//  ProviderManager.swift
+//  SeaEye
+//
+//  Created by Eoin Nolan on 28/03/2015.
+//  Copyright (c) 2015 Nolaneo. All rights reserved.
+//
+
+import Cocoa
+
+class ProviderManager {
+    func loadProjects() {
+        
+    }
+    
+    func runBuildUpdate() {
+        
+    }
+    
+}
