+++ SeaEye/models/providers/CircleCI/CircleCIBuild.swift
@@ -1,0 +1,9 @@
+//
+//  CircleCIBuild.swift
+//  SeaEye
+//
+//  Created by Eoin Nolan on 29/03/2015.
+//  Copyright (c) 2015 Nolaneo. All rights reserved.
+//
+
+import Foundation
