+++ SeaEye/models/providers/Codeship/CodeshipBuild.swift
@@ -1,0 +1,9 @@
+//
+//  CodeshipBuild.swift
+//  SeaEye
+//
+//  Created by Eoin Nolan on 29/03/2015.
+//  Copyright (c) 2015 Nolaneo. All rights reserved.
+//
+
+import Foundation
