+++ circle.yml
@@ -1,3 +1,3 @@
 test:
   override:
-     - xctool -scheme SeaEye -reporter pretty -reporter junit:$CIRCLE_TEST_REPORTS/xcode/results.xml build test: {retries: 2}
+    - xctool -scheme SeaEye -reporter pretty -reporter junit:$CIRCLE_TEST_REPORTS/xcode/results.xml build CIRCLE_TEST_REPORTS
